pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "engram-subnet"
version = "0.1.0"
description = "Decentralized IPFS-style vector database on Bittensor — the permanent semantic memory layer for AI."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["bittensor", "vector-database", "embeddings", "decentralized", "RAG", "semantic-search"]

dependencies = [
    "bittensor>=7.4.0",
    "qdrant-client>=1.9.0",
    "faiss-cpu>=1.8.0",
    "openai>=1.30.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
engram = "engram.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
